+++ noteable_magics/planar_ally_client/api.py
@@ -1,10 +1,13 @@
 from typing import Any, Dict
 
 import requests
+import structlog
 from urllib3.util import Timeout
 
-from .errors import PlanarAllyAPIError, PlanarAllyAPITimeoutError, PlanarAllyBadAPIResponseError
+from . import errors
 from .types import FileKind, RemoteStatus, UserMessage
+
+logger = structlog.get_logger(__name__)
 
 
 class PlanarAllyAPI:
@@ -20,6 +23,8 @@ class PlanarAllyAPI:
         self._timeout = Timeout(total=total_timeout_seconds, connect=0.5)
 
     def fs(self, kind: FileKind) -> "FileSystemAPI":
+        if kind is FileKind.dataset:
+            return DatasetFileSystemAPI(self, kind)
         return FileSystemAPI(self, kind)
 
     def post(self, endpoint: str, operation: str) -> Dict[str, Any]:
@@ -34,11 +39,19 @@ class PlanarAllyAPI:
     def _request(self, method: str, endpoint: str, operation: str, **kwargs) -> Dict[str, Any]:
         full_url = f"{self._base_url}{endpoint}"
         kwargs.setdefault("timeout", self._timeout)
+        logger.debug(
+            "making api request to planar-ally",
+            method=method,
+            endpoint=endpoint,
+            operation=operation,
+        )
 
         try:
             resp = self._session.request(method, full_url, **kwargs)
-        except requests.Timeout:
-            raise PlanarAllyAPITimeoutError(operation)
+        except requests.Timeout as e:
+            raise errors.PlanarAllyAPITimeoutError(operation) from e
+        except requests.ConnectionError as e:
+            raise errors.PlanarAllyUnableToConnectError() from e
 
         return self._check_response(resp, operation)
 
@@ -49,9 +62,9 @@ class PlanarAllyAPI:
             response = resp.text
 
         if resp.status_code != 200:
-            raise PlanarAllyAPIError(resp.status_code, response, operation)
+            raise errors.PlanarAllyAPIError(resp.status_code, response, operation)
         elif isinstance(response, (str, bytes)):
-            raise PlanarAllyBadAPIResponseError()
+            raise errors.PlanarAllyBadAPIResponseError()
 
         return response
 
@@ -81,3 +94,17 @@ class FileSystemAPI:
     def get_remote_status(self, path: str) -> RemoteStatus:
         resp = self._api.get(f"{self._url_prefix}/{path}/status", "get file status")
         return RemoteStatus.parse_obj(resp)
+
+
+class DatasetFileSystemAPI(FileSystemAPI):
+    def __init__(self, api: PlanarAllyAPI, kind: FileKind):
+        super().__init__(api, kind)
+
+    def delete(self, path: str) -> UserMessage:
+        raise errors.PlanarAllyError("delete is not supported for dataset files")
+
+    def move(self, path: str) -> UserMessage:
+        raise errors.PlanarAllyError("move is not supported for dataset files")
+
+    def get_remote_status(self, path: str) -> RemoteStatus:
+        raise errors.PlanarAllyError("get_remote_status is not supported for dataset files")